backend/app/core/cloud_tasks.py
from google.cloud import tasks_v2
import os
import json

# Initialize Cloud Tasks Client
# client = tasks_v2.CloudTasksClient()

def create_video_task(prompt: str, image: str, project_id: str, location: str, queue: str):
    """
    Creates a task for video generation.
    """
    # parent = client.queue_path(project_id, location, queue)
    
    # task = {
    #     "http_request": {
    #         "http_method": tasks_v2.HttpMethod.POST,
    #         "url": "https://your-worker-url/generate",  # The worker URL that actually calls Veo
    #         "headers": {"Content-Type": "application/json"},
    #         "body": json.dumps({"prompt": prompt, "image": image}).encode()
    #     }
    # }
    
    # response = client.create_task(request={"parent": parent, "task": task})
    # return response
    print(f"[MOCK] Created Cloud Task for prompt: {prompt[:20]}...")
    return "mock-task-id"
