backend/requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-aiplatform
google-cloud-storage
google-cloud-tasks
python-dotenv
pydantic-settings
sqlmodel
